api_basica/requirements.txt
requests==2.24.0
pytest==4.1.1
attrs==19.3.0
django==3.0.6
djangorestframework==3.11.0
python-decouple==3.1
psycopg2==2.8.5
psycopg2-binary==2.8.5